Move HR-related tools to their own folder, add print statement for BookPTO functionality, and add SILLY_MODE
