+++ tools/data/employee_pto_data.json
@@ -8,7 +8,7 @@
         "hrsAddedPerMonth": 8
       },
       {
-        "email": "laine.smith@awesome.com",
+        "email": "lainecaseysmith@gmail.com",
         "currentPTOHrs": 40,
         "hrsAddedPerMonth": 12
       }
